requirements.txt
Flask==0.10.1
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
Flask-RESTful==0.3.4
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Cors==2.1.2
PyJWT==1.4.0
PyMySQL==0.6.6
PyYAML==3.11
autopep8==1.2.1
inflect==0.2.5
marshmallow==2.3.0
marshmallow-jsonapi==0.3.0
uWSGI==2.0.11.2
celery==3.1.23
redis==2.10.5
suds-jurko==0.6